Static plugins
